# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de API, backend",
        "version": "0.1"
    },
    "paths": {
...
        "/ICECREAM/public/api/get-tiendas": {
            "get": {
                "tags": [
                    "Tiendas"
                ],
                "description": "Obtiene todas las tiendas",
                "operationId": "5e4723ca905dd1fb9270aa908ad01e19",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "nombre tienda"
                                            },
                                            "ubicacion_direccion": {
                                                "type": "string",
                                                "example": "calle tal"
                                            },
                                            "lat": {
                                                "type": "string",
                                                "example": "9.5644"
                                            },
                                            "lon": {
                                                "type": "string",
                                                "example": ".9231231"
                                            },
                                            "fecha_registro": {
                                                "type": "string",
                                                "example": "2024-02-29 09:27:54"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": [
                                            {
                                                "id": 1,
                                                "nombre": "nombre tienda",
                                                "ubicacion_direccion": "calle tal",
                                                "lat": "9.231321",
                                                "lon": "-9.56456",
                                                "fecha_registro": "2024-02-29 09:27:54"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario fuera de sesion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/ICECREAM/public/api/creaTienda": {
            "post": {
                "tags": [
                    "Tiendas"
                ],
                "description": "Guarda tiendas",
                "operationId": "ec3b4aa54a19f7ed2ae59c6cd4d5df6b",
                "requestBody": {
                    "description": "El cuerpo del body debe de tener la siguiente estructura",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "nombre": {
                                            "type": "string",
                                            "example": "nombre tienda"
                                        },
                                        "ubicacion_direccion": {
                                            "type": "string",
                                            "example": "ubicacion_direccion..."
                                        },
                                        "lat": {
                                            "type": "string",
                                            "example": "9.54654"
                                        },
                                        "lon": {
                                            "type": "string",
                                            "example": "-9.54654"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "examples": {
                                "result": {
                                    "summary": "",
                                    "value": {
                                        "nombre": "nombre tienda",
                                        "ubicacion_direccion": "calle tal, numero tal",
                                        "lat": "9.112121",
                                        "lon": "-9.12313"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": "..."
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "En proceso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": {
                                            "success": false,
                                            "message": "erro",
                                            "data": "Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/ICECREAM/public/api/editaTienda": {
            "put": {
                "tags": [
                    "Tiendas"
                ],
                "description": "Guarda tiendas",
                "operationId": "738b03a8601848e0aee90d0a16221528",
                "requestBody": {
                    "description": "El cuerpo del body debe de tener la siguiente estructura",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "nombre": {
                                            "type": "string",
                                            "example": "nombre tienda"
                                        },
                                        "ubicacion_direccion": {
                                            "type": "string",
                                            "example": "ubicacion_direccion..."
                                        },
                                        "lat": {
                                            "type": "string",
                                            "example": "9.54654"
                                        },
                                        "lon": {
                                            "type": "string",
                                            "example": "-9.54654"
                                        },
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "examples": {
                                "result": {
                                    "summary": "",
                                    "value": {
                                        "nombre": "nombre tienda",
                                        "ubicacion_direccion": "calle tal, numero tal",
                                        "lat": "9.112121",
                                        "lon": "-9.12313",
                                        "id": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": "..."
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "En proceso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "result": {
                                        "summary": "",
                                        "value": {
                                            "success": false,
                                            "message": "erro",
                                            "data": "Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}